Allow finding the Steam call from multiple providers
