+++ src/backend/providers/ProviderUtils.h
@@ -23,6 +23,7 @@
 
 namespace providers {
 
+#ifdef WITH_JSON_CACHE
 void cache_json(const QString& provider_prefix,
                 const QString& provider_dir,
                 const QString& entryname,
@@ -33,5 +34,11 @@ QJsonDocument read_json_from_cache(const QString& provider_prefix,
 void delete_cached_json(const QString& provider_prefix,
                         const QString& provider_dir,
                         const QString& entryname);
+#endif // WITH_JSON_CACHE
+
+#ifdef Q_OS_LINUX
+QString steam_flatpak_data_dir();
+#endif // Q_OS_LINUX
+QString find_steam_call();
 
 } // namespace providers
